requirements.txt
python-telegram-bot
python-dotenv
ffmpeg-python
